Add the third overworld stratum (blackstone)
This also includes a retexture/remodel of blackstone, using the BTW CE team's Jappa-style textures. Like with deepslate this is also extremely helpful to allow for the existing vanilla textures for the cracked/broken versions.

The bedrock has to be re-added thanks to the surface rule overriding it. Maybe there's a way to re-order things such that it isn't necessary?